# backend/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: BookFlow API
  version: 0.1.0
  summary: API for a private browser-based ebook reader
  description: Contract for the BookFlow modular monolith. All timestamps are RFC 3339 UTC unless a timezone parameter explicitly
    controls aggregation. Authentication can use the HttpOnly session cookie in browsers or a bearer access token for trusted
    clients. Error responses never expose SQL, stack traces, object-store keys, provider diagnostics, or secrets.
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
- url: /
  description: Same-origin API (frontend proxy in browser deployments)
security:
- cookieAuth: []
- bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      summary: Register a user
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '201':
          description: User and browser session created
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            X-CSRF-Token:
              $ref: '#/components/headers/CsrfToken'
            Set-Cookie:
              $ref: '#/components/headers/SetCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  email: reader@example.com
                  password: correct horse battery staple
                  display_name: Reader
                  locale: en
                  timezone: Asia/Yekaterinburg
                  device_key: browser-installation-uuid
                  device_name: Firefox on laptop
      security: []
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Log in
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Authenticated; refresh/session cookie is rotated
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            X-CSRF-Token:
              $ref: '#/components/headers/CsrfToken'
            Set-Cookie:
              $ref: '#/components/headers/SetCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  email: reader@example.com
                  password: correct horse battery staple
                  device_name: Firefox on laptop
      security: []
  /api/v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Rotate refresh token
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Session rotated; replacement cookies set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            X-CSRF-Token:
              $ref: '#/components/headers/CsrfToken'
            Set-Cookie:
              $ref: '#/components/headers/SetCookie'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: Consumes the current refresh token exactly once. Reuse revokes the affected token family.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      security: []
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Log out this session
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Session revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
  /api/v1/auth/logout-all:
    post:
      operationId: logoutAll
      summary: Log out all devices
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: All sessions and refresh tokens revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get current user
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Current user and device
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/devices:
    get:
      operationId: listDevices
      summary: List active devices
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Active and recently revoked user devices
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/devices/{deviceId}:
    delete:
      operationId: revokeDevice
      summary: Revoke a device and its sessions
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Device revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '400':
          $ref: '#/components/responses/ValidationError'
      description: A user cannot revoke a device owned by another user; associated refresh sessions are revoked transactionally.
...
components:
  schemas:
    ErrorCode:
      type: string
      pattern: ^[A-Z][A-Z0-9_]+$
      description: Stable machine-readable code. Clients must tolerate newly added codes.
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            details:
              description: Structured details or null.
            request_id:
              type: string
              format: uuid
          required:
          - code
          - message
          - details
          - request_id
      required:
      - error
      example:
        error:
          code: BOOK_NOT_FOUND
          message: Book was not found
          details: null
          request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        locale:
          type: string
          enum:
          - ru
          - en
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - email
      - display_name
      - locale
      - timezone
      - created_at
      - updated_at
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_key:
          type: string
        name:
          type: string
        user_agent:
          type: string
        last_seen_at:
          type: string
          format: date-time
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
      - id
      - device_key
      - name
      - user_agent
      - last_seen_at
      - created_at
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 1024
          writeOnly: true
        display_name:
          type: string
          maxLength: 100
        locale:
          type: string
          enum:
          - en
          - ru
          default: en
        timezone:
          type: string
          default: UTC
          description: IANA timezone name.
        device_key:
          type: string
          maxLength: 200
        device_name:
          type: string
          maxLength: 100
      required:
      - email
      - password
      - display_name
      additionalProperties: false
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 1024
          writeOnly: true
        device_key:
          type: string
          maxLength: 200
        device_name:
          type: string
          maxLength: 100
      required:
      - email
      - password
      additionalProperties: false
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          minLength: 32
          writeOnly: true
      description: Non-browser clients may submit an explicit refresh token. Browser cookie auth sends no JSON token and normally
        omits the body.
      additionalProperties: false
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        device:
          $ref: '#/components/schemas/Device'
        access_expires_at:
          type: string
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time
      required:
      - user
      - device
      - access_expires_at
      - refresh_expires_at
      description: Safe browser-auth response. Access and refresh token values are set only in HttpOnly cookies and are never
        returned in this JSON body.
    CurrentUserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
      - user
...
    DeviceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Device'
      required:
      - items
...
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-supplied UUID; the server validates or replaces it.
      schema:
        type: string
        format: uuid
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for unsafe requests authenticated by cookie; omit for bearer or explicit refresh-token requests.
        Must equal the signed readable CSRF cookie.
      schema:
        type: string
        minLength: 32
  headers:
    RequestId:
      description: Request identifier used in logs and traces.
      schema:
        type: string
        format: uuid
      example: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
    RetryAfter:
      description: Seconds until a retry is allowed.
      schema:
        type: integer
        minimum: 1
      example: 30
    CsrfToken:
      description: Signed double-submit token. Echo this value as X-CSRF-Token for unsafe cookie-authenticated requests.
      schema:
        type: string
    SetCookie:
      description: HttpOnly access/refresh cookies plus the readable signed CSRF cookie. Multiple Set-Cookie header fields
        are emitted.
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Request validation failed
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
    Unauthorized:
      description: Authentication is required or no longer valid
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication is required or no longer valid
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
    Forbidden:
      description: The authenticated user does not own this resource
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: The authenticated user does not own this resource
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
    NotFound:
      description: Resource was not found
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource was not found
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
    Conflict:
      description: The mutation conflicts with current state
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: CONFLICT
              message: The mutation conflicts with current state
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMITED
              message: Rate limit exceeded
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
...
    InternalError:
      description: An internal error occurred
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: An internal error occurred
              details: {}
              request_id: 0196d44a-0ee0-7530-9b4a-f4bea9e8e590
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bookflow_session
      description: HttpOnly, Secure in production, SameSite browser session cookie.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token for non-browser/trusted API clients.